[refactor] App: update package namespace
