+++ app/src/androidTest/java/com/rwmobi/giphytrending/CustomTestRunner.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2021. Ryan Wong (hello@rwmobi.co.uk)
+ *
+ */
+
+package com.rwmobi.giphytrending
+
+import android.app.Application
+import android.content.Context
+import androidx.test.runner.AndroidJUnitRunner
+import dagger.hilt.android.testing.HiltTestApplication
+
+class CustomTestRunner : AndroidJUnitRunner() {
+
+    override fun newApplication(cl: ClassLoader?, name: String?, context: Context?): Application {
+        return super.newApplication(cl, HiltTestApplication::class.java.name, context)
+    }
+}
