+++ app/src/androidTest/java/com/rwmobi/giphytrending/data/source/local/TrendingEntityTestData.kt
@@ -1,0 +1,102 @@
+package com.rwmobi.giphytrending.data.source.local
+
+import java.util.Date
+
+object TrendingEntityTestData {
+    val case1 by lazy {
+        TrendingEntity(
+            id = "5KF7hci72bv2ZbchuT",
+            previewUrl = "https://media2.giphy.com/media/5KF7hci72bv2ZbchuT/200w.gif",
+            imageUrl = "https://media2.giphy.com/media/5KF7hci72bv2ZbchuT/giphy.gif",
+            webUrl = "https://giphy.com/gifs/studiosoriginals-cooking-chef-cook-5KF7hci72bv2ZbchuT",
+            title = "Chef Cooking GIF by GIPHY Studios Originals",
+            type = "gif",
+            username = "studiosoriginals",
+            trendingDateTime = Date(1636679711000),
+            importDateTime = Date(1635978358000),
+        )
+    }
+
+    val case2 by lazy {
+        TrendingEntity(
+            id = "uaIAIw3ELuk69mhZ5I",
+            previewUrl = "https://media3.giphy.com/media/uaIAIw3ELuk69mhZ5I/200w.gif",
+            imageUrl = "https://media3.giphy.com/media/uaIAIw3ELuk69mhZ5I/giphy.gif",
+            webUrl = "https://giphy.com/gifs/creative-courage-vidhyan-as-a-nation-we-have-lot-of-obligations-no-obligation-uaIAIw3ELuk69mhZ5I",
+            title = "Joe Biden GIF by Creative Courage",
+            type = "gif",
+            username = "creative-courage",
+            trendingDateTime = Date(-62170156725000),
+            importDateTime = Date(1605041010000),
+        )
+    }
+    val case2Modified by lazy {
+        TrendingEntity(
+            id = "uaIAIw3ELuk69mhZ5I",
+            previewUrl = "https://media3.giphy.com/media/3o84U4VlzJveib6Lcs/200w.gif",
+            imageUrl = "https://media3.giphy.com/media/3o84U4VlzJveib6Lcs/giphy.gif",
+            webUrl = "https://giphy.com/gifs/toferra-ouch-3o84U4VlzJveib6Lcs",
+            title = "Cam Newton Hit GIF by Tomi Ferraro, Sportz",
+            type = "gif",
+            username = "toferra",
+            trendingDateTime = Date(1487871000000),
+            importDateTime = Date(1487870240000),
+        )
+    }
+
+    val case3 by lazy {
+        TrendingEntity(
+            "etKSrsbbKbqwW6vzOg",
+            previewUrl = "https://media1.giphy.com/media/etKSrsbbKbqwW6vzOg/200w.gif",
+            imageUrl = "https://media1.giphy.com/media/etKSrsbbKbqwW6vzOg/giphy.gif",
+            webUrl = "https://giphy.com/gifs/studiosoriginals-chicken-dinner-winner-etKSrsbbKbqwW6vzOg",
+            title = "Winner Winner Win GIF by GIPHY Studios Originals",
+            type = "gif",
+            username = "studiosoriginals",
+            trendingDateTime = Date(1636678809000),
+            importDateTime = Date(1636417528000),
+        )
+    }
+
+    val case4 by lazy {
+        TrendingEntity(
+            id = "SwIMZUJE3ZPpHAfTC4",
+            previewUrl = "https://media1.giphy.com/media/SwIMZUJE3ZPpHAfTC4/200w.gif",
+            imageUrl = "https://media1.giphy.com/media/SwIMZUJE3ZPpHAfTC4/giphy.gif",
+            webUrl = "https://giphy.com/gifs/mumbaiindians-happy-birthday-mi-ipl-SwIMZUJE3ZPpHAfTC4",
+            title = "Happy Birthday GIF by Mumbai Indians",
+            type = "gif",
+            username = "mumbaiindians",
+            trendingDateTime = Date(-62170156725000),
+            importDateTime = Date(1581493566000),
+        )
+    }
+
+    val case5 by lazy {
+        TrendingEntity(
+            id = "NEvPzZ8bd1V4Y",
+            previewUrl = "https://media4.giphy.com/media/NEvPzZ8bd1V4Y/200w.gif",
+            imageUrl = "https://media4.giphy.com/media/NEvPzZ8bd1V4Y/giphy.gif",
+            webUrl = "https://giphy.com/gifs/editingandlayout-NEvPzZ8bd1V4Y",
+            title = "Proud Of You Reaction GIF",
+            type = "gif",
+            username = "editingandlayout",
+            trendingDateTime = Date(1621233914000),
+            importDateTime = Date(1410285308000),
+        )
+    }
+
+    val case6 by lazy {
+        TrendingEntity(
+            id = "3o84U4VlzJveib6Lcs",
+            previewUrl = "https://media3.giphy.com/media/3o84U4VlzJveib6Lcs/200w.gif",
+            imageUrl = "https://media3.giphy.com/media/3o84U4VlzJveib6Lcs/giphy.gif",
+            webUrl = "https://giphy.com/gifs/toferra-ouch-3o84U4VlzJveib6Lcs",
+            title = "Cam Newton Hit GIF by Tomi Ferraro, Sportz",
+            type = "gif",
+            username = "toferra",
+            trendingDateTime = Date(1487871000000),
+            importDateTime = Date(1487870240000),
+        )
+    }
+}
