+++ app/src/debug/res/values/strings.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright (c) 2021. Ryan Wong (hello@ryanwong.co.uk)
+  ~ Copyright (c) 2021. Ryan Wong (hello@rwmobi.co.uk)
   ~
   -->
