--- app/src/testFixtures/java/uk/ryanwong/giphytrending/data.repository/MockGiphyRepository.kt
@@ -1,16 +1,0 @@
-package uk.ryanwong.giphytrending.data.repository
-
-import uk.ryanwong.giphytrending.domain.model.GiphyImageItemDomainModel
-import uk.ryanwong.giphytrending.domain.repository.GiphyRepository
-
-class MockGiphyRepository : GiphyRepository {
-    var mockFetchCachedTrendingResult: Result<List<GiphyImageItemDomainModel>>? = null
-    override suspend fun fetchCachedTrending(): Result<List<GiphyImageItemDomainModel>> {
-        return mockFetchCachedTrendingResult ?: Result.success(emptyList())
-    }
-
-    var mockReloadTrendingResult: Result<List<GiphyImageItemDomainModel>>? = null
-    override suspend fun reloadTrending(apiMaxEntries: Int): Result<List<GiphyImageItemDomainModel>> {
-        return mockReloadTrendingResult ?: Result.success(emptyList())
-    }
-}
